Split entire SDL2 WindowsMouseHandler implementation
