--- osu.Framework/Platform/Windows/WindowsMouseHandler_SDL3.cs
@@ -1,13 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Framework.Platform.Windows
-{
-    internal partial class WindowsMouseHandler
-    {
-        private bool bindHandlerSDL3(GameHost host)
-        {
-            return false;
-        }
-    }
-}
